+++ crates/propaga-propagators/src/lib.rs
@@ -27,6 +27,7 @@ mod float_times;
 mod forbidden_assignment;
 mod gcc;
 mod int_abs;
+mod int_div;
 mod int_minmax;
 mod int_times;
 mod inverse;
@@ -85,6 +86,7 @@ pub use forbidden_assignment::{
 };
 pub use gcc::{CardinalityBound, GlobalCardinalityPropagator};
 pub use int_abs::IntAbsPropagator;
+pub use int_div::IntDivPropagator;
 pub use int_minmax::{IntMinMaxOp, IntMinMaxPropagator};
 pub use int_times::IntTimesPropagator;
 pub use inverse::InversePropagator;